Hier werden die Punkte beim Aufruf gezeichnet, aber nicht nicht skaliert dargestellt, ich werde versuchen an der skalierung zu arbeiten..

+++ Ameisen/src/Daten_einlesen.java
@@ -2,10 +2,10 @@
 public class Daten_einlesen {
 
 	private int id;
-	private double xPos;
-	private double yPos;
+	private int xPos;
+	private int yPos;
 	
-	public Daten_einlesen(int id, double xPos, double yPos) {
+	public Daten_einlesen(int id, int xPos, int yPos) {
 		super();
 		this.id = id;
 		this.xPos = xPos;
@@ -17,16 +17,16 @@ public class Daten_einlesen {
 	public void setId(int id) {
 		this.id = id;
 	}
-	public double getxPos() {
+	public int getxPos() {
 		return xPos;
 	}
-	public void setxPos(double xPos) {
+	public void setxPos(int xPos) {
 		this.xPos = xPos;
 	}
-	public double getyPos() {
+	public int getyPos() {
 		return yPos;
 	}
-	public void setyPos(double yPos) {
+	public void setyPos(int yPos) {
 		this.yPos = yPos;
 	}
 	@Override
